% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_coord.R
\name{convertir_coord}
\alias{convertir_coord}
\title{Convertir formato de coordenadas}
\usage{
convertir_coord(x, y, output = "lonlat")
}
\arguments{
\item{x}{Coordenada \code{x}}

\item{y}{Coordenada \code{y}}

\item{output}{Formato de coordenada de salida. Uno de \code{gkba}, \code{lonlat} o \code{degrees}.}
}
\description{
Convierte coordenadas a formatos \code{GKBA}, \code{WGS84 (Lon/Lat)} y
\verb{WGS84 (grados, minutos y segundos)}. El parámetro \code{output} puede ser uno de
\code{gkba}, \code{lonlat} o \code{degrees}.
Acepta cualquier formato de entrada. El formato de entrada es
detectado automáticamente.
}
\examples{
convertir_coord(x = -58.37583628694437,
                y = -34.59857730467378,
                output = "degrees")
}
