% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode2.R
\name{geocode2}
\alias{geocode2}
\title{Buscar cordenadas a partir de dirección (Geocoding) basado en USIG 2.1.2}
\usage{
geocode2(direccion, maxOptions = 1, output = "lonlat", srid = "4326")
}
\arguments{
\item{direccion}{Character. Incluye calle y altura o intersección.}

\item{maxOptions}{Numeric. Cantidad máxima de opciones a devolver}

\item{output}{Uno de \code{gkba}, \code{lonlat} o \code{degrees}.
Formato de las coordenadas de salida.}

\item{srid}{Numeric or string. SRID a usar en la geocodificación. El
default es \code{srid = 4326}}
}
\description{
Encuentra la coordenada más cercana a la dirección provista. Formato de
output es \code{lonlat} (longitud, latitud) por default. Seleccionar \code{gkba} o
\code{degrees} en el parámetro \code{output} para modificar.
A diferencia de \code{geocode} (que utiliza el sistema USIG 2.2),
se utiliza el sistema previo USIG 2.1.2.
}
\examples{
geocode2(calle  = "Córdoba av 637, caba")

geocode(calle  = "Córdoba y florida, caba")
}
