% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar_direccion.R
\name{normalizar_direccion}
\alias{normalizar_direccion}
\title{Normalizador de direcciones}
\usage{
normalizar_direccion(calle, altura = NULL, desambiguar = 1)
}
\arguments{
\item{calle}{Nombre de la calle o intersección completa}

\item{altura}{Altura de la calle (vacío en caos de intersección)}

\item{desambiguar}{Permitir permutaciones de las palabras de \code{calle} (1)
o no (0)}
}
\description{
Utiliza el servicio USIG para normalizar una dirección. Devuelve el nombre
de la calle normalizado y el código identificador. Se puede buscar una
calle y altura o una intersección, en cuyo caso la intersección se incluye
en la variable \code{calle}.
}
\examples{
normalizar_direccion("córdoba", 637)
}
