% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Buscar direcci칩n a partir de coordenadas (Reverse Geocoding)}
\usage{
reverse_geocode(x, y)
}
\arguments{
\item{x}{Coordenada x o longitud}

\item{y}{Coordenada y o latitud}
}
\description{
Encuentra la direcci칩n m치s cercana a la coordenada provista. Formato de
coordenada detectado autom치ticamente.
}
\examples{
coord_a_dir(x = -58.37583628694437,
            y = -34.59857730467378)
}
